ext-costs moved to logic config extracted
