+++ runtime/near-runtime-fees/src/lib.rs
@@ -41,8 +41,6 @@ impl Fee {
 
 #[derive(Debug, Serialize, Deserialize, Clone, Hash, PartialEq, Eq)]
 pub struct RuntimeFeesConfig {
-    /// Costs for runtime externals
-    pub ext_costs: ExtCostsConfig,
     /// Describes the cost of creating an action receipt, `ActionReceipt`, excluding the actual cost
     /// of actions.
     pub action_receipt_creation_config: Fee,
@@ -57,88 +55,6 @@ pub struct RuntimeFeesConfig {
     pub burnt_gas_reward: Fraction,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, Hash, PartialEq, Eq)]
-pub struct ExtCostsConfig {
-    pub input_byte: Gas,
-    /// Storage trie read key base cost
-    pub storage_read_base: Gas,
-    /// Storage trie read key per byte cost
-    pub storage_read_key_byte: Gas,
-    /// Storage trie read value cost per byte cost
-    pub storage_read_value_byte: Gas,
-    /// Storage trie write key base cost
-    pub storage_write_base: Gas,
-    /// Storage trie write key per byte cost
-    pub storage_write_key_byte: Gas,
-    /// Storage trie write value per byte cost
-    pub storage_write_value_byte: Gas,
-    /// Storage trie check for key existence cost base
-    pub storage_has_key_base: Gas,
-    /// Storage trie check for key existence per key byte
-    pub storage_has_key_byte: Gas,
-    /// Remove key from trie base cost
-    pub storage_remove_base: Gas,
-    /// Remove key from trie per byte cost
-    pub storage_remove_key_byte: Gas,
-    /// Remove key from trie ret value byte cost
-    pub storage_remove_ret_value_byte: Gas,
-    /// Create trie prefix iterator cost base
-    pub storage_iter_create_prefix_base: Gas,
-    /// Create trie range iterator cost base
-    pub storage_iter_create_range_base: Gas,
-    /// Create trie iterator per key byte cost
-    pub storage_iter_create_key_byte: Gas,
-    /// Trie iterator per key base cost
-    pub storage_iter_next_base: Gas,
-    /// Trie iterator next key byte cost
-    pub storage_iter_next_key_byte: Gas,
-    /// Trie iterator next key byte cost
-    pub storage_iter_next_value_byte: Gas,
-    /// Base cost for reading from register
-    pub read_register_base: Gas,
-    /// Base cost for reading byte from register
-    pub read_register_byte: Gas,
-    /// Base cost for writing into register
-    pub write_register_base: Gas,
-    /// Base cost for writing byte into register
-    pub write_register_byte: Gas,
-    /// Get account balance cost
-    pub account_balance: Gas,
-    /// Get prepaid gas cost
-    pub prepaid_gas: Gas,
-    /// Get used gas cost
-    pub used_gas: Gas,
-    /// Cost of getting random seed
-    pub random_seed: Gas,
-    /// Cost of getting sha256
-    pub sha256: Gas,
-    /// Cost of getting sha256
-    pub sha256_byte: Gas,
-    /// Get account attached_deposit base cost
-    pub attached_deposit: Gas,
-    /// Get storage usage cost
-    pub storage_usage: Gas,
-    /// Get a current block height base cost
-    pub block_index: Gas,
-    /// Get a current timestamp base cost
-    pub block_timestamp: Gas,
-    /// Cost for getting a current account base
-    pub current_account_id: Gas,
-    /// Cost for getting a current account per byte
-    pub current_account_id_byte: Gas,
-    /// Cost for getting a signer account id base
-    pub signer_account_id: Gas,
-    /// Cost for getting a signer account per byte
-    pub signer_account_id_byte: Gas,
-    /// Cost for getting a signer public key
-    pub signer_account_pk: Gas,
-    /// Cost for getting a signer public key per byte
-    pub signer_account_pk_byte: Gas,
-    /// Cost for getting a predecessor account
-    pub predecessor_account_id: Gas,
-    /// Cost for getting a predecessor account per byte
-    pub predecessor_account_id_byte: Gas,
-}
 
 /// Describes the cost of creating a data receipt, `DataReceipt`.
 #[derive(Debug, Serialize, Deserialize, Clone, Hash, PartialEq, Eq)]
@@ -209,48 +125,6 @@ pub struct StorageUsageConfig {
 impl Default for RuntimeFeesConfig {
     fn default() -> Self {
         Self {
-            ext_costs: ExtCostsConfig {
-                input_byte: 1,
-                storage_read_base: 1,
-                storage_read_key_byte: 1,
-                storage_read_value_byte: 1,
-                storage_write_base: 1,
-                storage_write_key_byte: 1,
-                storage_write_value_byte: 1,
-                storage_has_key_base: 1,
-                storage_has_key_byte: 1,
-                storage_remove_base: 1,
-                storage_remove_key_byte: 1,
-                storage_remove_ret_value_byte: 1,
-                storage_iter_create_prefix_base: 1,
-                storage_iter_create_range_base: 1,
-                storage_iter_create_key_byte: 1,
-                storage_iter_next_base: 1,
-                storage_iter_next_key_byte: 1,
-                storage_iter_next_value_byte: 1,
-                read_register_base: 1,
-                read_register_byte: 1,
-                write_register_base: 1,
-                write_register_byte: 1,
-                account_balance: 1,
-                prepaid_gas: 1,
-                used_gas: 1,
-                random_seed: 1,
-                sha256: 1,
-                sha256_byte: 1,
-                attached_deposit: 1,
-                storage_usage: 1,
-                block_index: 1,
-                block_timestamp: 1,
-                current_account_id: 1,
-                current_account_id_byte: 1,
-                signer_account_id: 1,
-                signer_account_id_byte: 1,
-                signer_account_pk: 1,
-                signer_account_pk_byte: 1,
-                predecessor_account_id: 1,
-                predecessor_account_id_byte: 1,
-            },
             action_receipt_creation_config: Fee { send_sir: 10, send_not_sir: 10, execution: 10 },
             data_receipt_creation_config: DataReceiptCreationConfig {
                 base_cost: Fee { send_sir: 10, send_not_sir: 10, execution: 10 },
@@ -296,48 +170,6 @@ impl RuntimeFeesConfig {
     pub fn free() -> Self {
         let free = Fee { send_sir: 0, send_not_sir: 0, execution: 0 };
         RuntimeFeesConfig {
-            ext_costs: ExtCostsConfig {
-                input_byte: 0,
-                storage_read_base: 0,
-                storage_read_key_byte: 0,
-                storage_read_value_byte: 0,
-                storage_write_base: 0,
-                storage_write_key_byte: 0,
-                storage_write_value_byte: 0,
-                storage_has_key_base: 0,
-                storage_has_key_byte: 0,
-                storage_remove_base: 0,
-                storage_remove_key_byte: 0,
-                storage_remove_ret_value_byte: 0,
-                storage_iter_create_prefix_base: 0,
-                storage_iter_create_range_base: 0,
-                storage_iter_create_key_byte: 0,
-                storage_iter_next_base: 0,
-                storage_iter_next_key_byte: 0,
-                storage_iter_next_value_byte: 0,
-                read_register_base: 0,
-                read_register_byte: 0,
-                write_register_base: 0,
-                write_register_byte: 0,
-                account_balance: 0,
-                prepaid_gas: 0,
-                used_gas: 0,
-                random_seed: 0,
-                sha256: 0,
-                sha256_byte: 0,
-                attached_deposit: 0,
-                storage_usage: 0,
-                block_index: 0,
-                block_timestamp: 0,
-                current_account_id: 0,
-                current_account_id_byte: 0,
-                signer_account_id: 0,
-                signer_account_id_byte: 0,
-                signer_account_pk: 0,
-                signer_account_pk_byte: 0,
-                predecessor_account_id: 0,
-                predecessor_account_id_byte: 0,
-            },
             action_receipt_creation_config: free.clone(),
             data_receipt_creation_config: DataReceiptCreationConfig {
                 base_cost: free.clone(),
