+++ runtime/near-vm-logic/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "near-vm-logic"
-version = "0.3.0"
+version = "0.4.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2018"
 license = "Apache-2.0"
@@ -16,7 +16,7 @@ This crate implements the specification of the interface that Near blockchain ex
 bs58 = "0.2.2"
 sodiumoxide = { version = "0.2.2", optional = true }
 serde = { version = "1.0", features = ["derive"] }
-near-runtime-fees = { path = "../near-runtime-fees", version = "0.3.0" }
+near-runtime-fees = { path = "../near-runtime-fees", version = "0.4.0" }
 near-vm-errors = { path = "../near-vm-errors" }
 
 [dev-dependencies]
