+++ runtime/near-vm-logic/src/logic.rs
@@ -7,7 +7,7 @@ use crate::types::{
     StorageUsage,
 };
 use crate::{HostError, HostErrorOrStorageError};
-use near_runtime_fees::ExtCostsConfig;
+use near_runtime_fees::RuntimeFeesConfig;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::mem::size_of;
@@ -20,8 +20,10 @@ pub struct VMLogic<'a> {
     ext: &'a mut dyn External,
     /// Part of Context API and Economics API that was extracted from the receipt.
     context: VMContext,
-    /// Parameters of Wasm and economic parameters.
+    /// Parameters for Wasm VM and costs of external functions execution
     config: &'a Config,
+    /// Fees for creating (async) actions on runtime.
+    fees_config: &'a RuntimeFeesConfig,
     /// If this method execution is invoked directly as a callback by one or more contract calls the
     /// results of the methods that made the callback are stored in this collection.
     promise_results: &'a [PromiseResult],
@@ -75,6 +77,7 @@ impl<'a> VMLogic<'a> {
         ext: &'a mut dyn External,
         context: VMContext,
         config: &'a Config,
+        fees_config: &'a RuntimeFeesConfig,
         promise_results: &'a [PromiseResult],
         memory: &'a mut dyn MemoryLike,
     ) -> Self {
@@ -86,6 +89,7 @@ impl<'a> VMLogic<'a> {
             ext,
             context,
             config,
+            fees_config,
             promise_results,
             memory,
             current_account_balance,
@@ -189,9 +193,12 @@ impl<'a> VMLogic<'a> {
     /// wrong location this function will overwrite memory that it is not supposed to overwrite causing an undefined behavior.
     pub fn read_register(&mut self, register_id: u64, ptr: u64) -> Result<()> {
         let Self { registers, memory, config, .. } = self;
-        self.gas_counter.pay_base(config.runtime_fees.ext_costs.read_register_base)?;
+        self.gas_counter.pay_base(config.ext_costs.read_register_base)?;
         let register = registers.get(&register_id).ok_or(HostError::InvalidRegisterId)?;
-        self.gas_counter.pay_per_byte(config.runtime_fees.ext_costs.read_register_byte, register.len() as u64)?;
+        self.gas_counter.pay_per_byte(
+            config.ext_costs.read_register_byte,
+            register.len() as u64,
+        )?;
         Self::memory_set(*memory, ptr, register)
     }
 
@@ -226,9 +233,9 @@ impl<'a> VMLogic<'a> {
         register_id: u64,
         data: &[u8],
     ) -> Result<()> {
-        gas_counter.pay_base(config.runtime_fees.ext_costs.write_register_base)?;
+        gas_counter.pay_base(config.ext_costs.write_register_base)?;
         gas_counter
-            .pay_per_byte(config.runtime_fees.ext_costs.write_register_byte, data.len() as u64)?;
+            .pay_per_byte(config.ext_costs.write_register_byte, data.len() as u64)?;
         if data.len() as u64 > config.max_register_size
             || registers.len() as u64 == config.max_number_registers
         {
@@ -260,10 +267,10 @@ impl<'a> VMLogic<'a> {
     /// If the registers exceed the memory limit returns `MemoryAccessViolation`.
     pub fn current_account_id(&mut self, register_id: u64) -> Result<()> {
         let Self { context, registers, gas_counter, config, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.current_account_id)?;
+        gas_counter.pay_base(config.ext_costs.current_account_id)?;
         let data = context.current_account_id.as_bytes();
         gas_counter.pay_per_byte(
-            config.runtime_fees.ext_costs.signer_account_id_byte,
+            config.ext_costs.signer_account_id_byte,
             data.len() as u64,
         )?;
         Self::internal_write_register(registers, gas_counter, config, register_id, data)
@@ -280,13 +287,13 @@ impl<'a> VMLogic<'a> {
     /// * If called as view function returns `ProhibitedInView`.
     pub fn signer_account_id(&mut self, register_id: u64) -> Result<()> {
         let Self { context, registers, gas_counter, config, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.signer_account_id)?;
+        gas_counter.pay_base(config.ext_costs.signer_account_id)?;
         if context.is_view {
             return Err(HostError::ProhibitedInView("signer_account_id".to_string()).into());
         }
         let data = context.signer_account_id.as_bytes();
         gas_counter.pay_per_byte(
-            config.runtime_fees.ext_costs.signer_account_id_byte,
+            config.ext_costs.signer_account_id_byte,
             data.len() as u64,
         )?;
         Self::internal_write_register(registers, gas_counter, config, register_id, data)
@@ -302,13 +309,13 @@ impl<'a> VMLogic<'a> {
     /// * If called as view function returns `ProhibitedInView`.
     pub fn signer_account_pk(&mut self, register_id: u64) -> Result<()> {
         let Self { context, registers, gas_counter, config, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.signer_account_id)?;
+        gas_counter.pay_base(config.ext_costs.signer_account_id)?;
         if context.is_view {
             return Err(HostError::ProhibitedInView("signer_account_pk".to_string()).into());
         }
         let data = context.signer_account_pk.as_slice();
         gas_counter.pay_per_byte(
-            config.runtime_fees.ext_costs.signer_account_pk_byte,
+            config.ext_costs.signer_account_pk_byte,
             data.len() as u64,
         )?;
         Self::internal_write_register(registers, gas_counter, config, register_id, data)
@@ -324,13 +331,13 @@ impl<'a> VMLogic<'a> {
     /// * If called as view function returns `ProhibitedInView`.
     pub fn predecessor_account_id(&mut self, register_id: u64) -> Result<()> {
         let Self { context, registers, gas_counter, config, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.predecessor_account_id)?;
+        gas_counter.pay_base(config.ext_costs.predecessor_account_id)?;
         if context.is_view {
             return Err(HostError::ProhibitedInView("predecessor_account_id".to_string()).into());
         }
         let data = context.predecessor_account_id.as_bytes();
         gas_counter.pay_per_byte(
-            config.runtime_fees.ext_costs.predecessor_account_id_byte,
+            config.ext_costs.predecessor_account_id_byte,
             data.len() as u64,
         )?;
         Self::internal_write_register(registers, gas_counter, config, register_id, data)
@@ -346,13 +353,13 @@ impl<'a> VMLogic<'a> {
 
     /// Returns the current block index.
     pub fn block_index(&mut self) -> Result<u64> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.block_index)?;
+        self.gas_counter.pay_base(self.config.ext_costs.block_index)?;
         Ok(self.context.block_index)
     }
 
     /// Returns the current block timestamp.
     pub fn block_timestamp(&mut self) -> Result<u64> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.block_timestamp)?;
+        self.gas_counter.pay_base(self.config.ext_costs.block_timestamp)?;
         Ok(self.context.block_timestamp)
     }
 
@@ -364,7 +371,7 @@ impl<'a> VMLogic<'a> {
     /// * A small fixed overhead for account metadata.
     ///
     pub fn storage_usage(&mut self) -> Result<StorageUsage> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.storage_usage)?;
+        self.gas_counter.pay_base(self.config.ext_costs.storage_usage)?;
         Ok(self.current_storage_usage)
     }
 
@@ -375,7 +382,7 @@ impl<'a> VMLogic<'a> {
     /// The current balance of the given account. This includes the attached_deposit that was
     /// attached to the transaction.
     pub fn account_balance(&mut self, balance_ptr: u64) -> Result<()> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.account_balance)?;
+        self.gas_counter.pay_base(self.config.ext_costs.account_balance)?;
         Self::memory_set(self.memory, balance_ptr, &self.current_account_balance.to_le_bytes())
     }
 
@@ -386,7 +393,7 @@ impl<'a> VMLogic<'a> {
     ///
     /// If called as view function returns `ProhibitedInView``.
     pub fn attached_deposit(&mut self, balance_ptr: u64) -> Result<()> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.attached_deposit)?;
+        self.gas_counter.pay_base(self.config.ext_costs.attached_deposit)?;
         if self.context.is_view {
             return Err(HostError::ProhibitedInView("attached_deposit".to_string()).into());
         }
@@ -399,7 +406,7 @@ impl<'a> VMLogic<'a> {
     ///
     /// If called as view function returns `ProhibitedInView`.
     pub fn prepaid_gas(&mut self) -> Result<Gas> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.prepaid_gas)?;
+        self.gas_counter.pay_base(self.config.ext_costs.prepaid_gas)?;
         if self.context.is_view {
             return Err(HostError::ProhibitedInView("prepaid_gas".to_string()).into());
         }
@@ -412,7 +419,7 @@ impl<'a> VMLogic<'a> {
     ///
     /// If called as view function returns `ProhibitedInView`.
     pub fn used_gas(&mut self) -> Result<Gas> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.used_gas)?;
+        self.gas_counter.pay_base(self.config.ext_costs.used_gas)?;
         if self.context.is_view {
             return Err(HostError::ProhibitedInView("used_gas".to_string()).into());
         }
@@ -430,7 +437,7 @@ impl<'a> VMLogic<'a> {
     /// If the size of the registers exceed the set limit `MemoryAccessViolation`.
     pub fn random_seed(&mut self, register_id: u64) -> Result<()> {
         let Self { context, registers, gas_counter, config, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.random_seed)?;
+        gas_counter.pay_base(config.ext_costs.random_seed)?;
         Self::internal_write_register(
             registers,
             gas_counter,
@@ -448,8 +455,8 @@ impl<'a> VMLogic<'a> {
     /// the limit with `MemoryAccessViolation`.
     pub fn sha256(&mut self, value_len: u64, value_ptr: u64, register_id: u64) -> Result<()> {
         let Self { memory, registers, gas_counter, config, ext, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.sha256)?;
-        gas_counter.pay_per_byte(config.runtime_fees.ext_costs.sha256_byte, value_len)?;
+        gas_counter.pay_base(config.ext_costs.sha256)?;
+        gas_counter.pay_per_byte(config.ext_costs.sha256_byte, value_len)?;
         let value = Self::memory_get(*memory, value_ptr, value_len)?;
         let value_hash = ext.sha256(&value)?;
         Self::internal_write_register(registers, gas_counter, config, register_id, &value_hash)
@@ -476,17 +483,17 @@ impl<'a> VMLogic<'a> {
     /// * `data_dependencies`: other contracts that this execution will be waiting on (or rather
     ///   their data receipts), where bool indicates whether this is sender=receiver communication.
     fn pay_gas_for_new_receipt(&mut self, sir: bool, data_dependencies: &[bool]) -> Result<()> {
-        let runtime_fees_cfg = &self.config.runtime_fees;
-        let mut use_gas = runtime_fees_cfg
+        let fees_config_cfg = &self.fees_config;
+        let mut use_gas = fees_config_cfg
             .action_receipt_creation_config
             .send_fee(sir)
-            .checked_add(runtime_fees_cfg.action_receipt_creation_config.exec_fee())
+            .checked_add(fees_config_cfg.action_receipt_creation_config.exec_fee())
             .ok_or(HostError::IntegerOverflow)?;
         for dep in data_dependencies {
             use_gas = use_gas
-                .checked_add(runtime_fees_cfg.data_receipt_creation_config.base_cost.send_fee(*dep))
+                .checked_add(fees_config_cfg.data_receipt_creation_config.base_cost.send_fee(*dep))
                 .ok_or(HostError::IntegerOverflow)?
-                .checked_add(runtime_fees_cfg.data_receipt_creation_config.base_cost.exec_fee())
+                .checked_add(fees_config_cfg.data_receipt_creation_config.base_cost.exec_fee())
                 .ok_or(HostError::IntegerOverflow)?;
         }
         self.gas_counter.deduct_gas(0, use_gas)
@@ -752,7 +759,7 @@ impl<'a> VMLogic<'a> {
         let (receipt_idx, sir) = self.promise_idx_to_receipt_idx_with_sir(promise_idx)?;
 
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.create_account_cost,
+            &self.fees_config.action_creation_config.create_account_cost,
             sir,
         )?;
 
@@ -789,11 +796,11 @@ impl<'a> VMLogic<'a> {
 
         let num_bytes = code.len() as u64;
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.deploy_contract_cost,
+            &self.fees_config.action_creation_config.deploy_contract_cost,
             sir,
         )?;
         self.gas_counter.pay_per_byte_gas_fee(
-            &self.config.runtime_fees.action_creation_config.deploy_contract_cost_per_byte,
+            &self.fees_config.action_creation_config.deploy_contract_cost_per_byte,
             num_bytes,
             sir,
         )?;
@@ -841,11 +848,11 @@ impl<'a> VMLogic<'a> {
 
         let num_bytes = (method_name.len() + arguments.len()) as u64;
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.function_call_cost,
+            &self.fees_config.action_creation_config.function_call_cost,
             sir,
         )?;
         self.gas_counter.pay_per_byte_gas_fee(
-            &self.config.runtime_fees.action_creation_config.function_call_cost_per_byte,
+            &self.fees_config.action_creation_config.function_call_cost_per_byte,
             num_bytes,
             sir,
         )?;
@@ -884,7 +891,7 @@ impl<'a> VMLogic<'a> {
         let (receipt_idx, sir) = self.promise_idx_to_receipt_idx_with_sir(promise_idx)?;
 
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.transfer_cost,
+            &self.fees_config.action_creation_config.transfer_cost,
             sir,
         )?;
 
@@ -924,7 +931,7 @@ impl<'a> VMLogic<'a> {
         let (receipt_idx, sir) = self.promise_idx_to_receipt_idx_with_sir(promise_idx)?;
 
         self.gas_counter
-            .pay_base_gas_fee(&self.config.runtime_fees.action_creation_config.stake_cost, sir)?;
+            .pay_base_gas_fee(&self.fees_config.action_creation_config.stake_cost, sir)?;
 
         self.deduct_balance(amount)?;
 
@@ -962,7 +969,7 @@ impl<'a> VMLogic<'a> {
         let (receipt_idx, sir) = self.promise_idx_to_receipt_idx_with_sir(promise_idx)?;
 
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.add_key_cost.full_access_cost,
+            &self.fees_config.action_creation_config.add_key_cost.full_access_cost,
             sir,
         )?;
 
@@ -1024,13 +1031,12 @@ impl<'a> VMLogic<'a> {
         // +1 is to account for null-terminating characters.
         let num_bytes = method_names.iter().map(|v| v.len() as u64 + 1).sum::<u64>();
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.add_key_cost.function_call_cost,
+            &self.fees_config.action_creation_config.add_key_cost.function_call_cost,
             sir,
         )?;
         self.gas_counter.pay_per_byte_gas_fee(
             &self
-                .config
-                .runtime_fees
+                .fees_config
                 .action_creation_config
                 .add_key_cost
                 .function_call_cost_per_byte,
@@ -1077,7 +1083,7 @@ impl<'a> VMLogic<'a> {
         let (receipt_idx, sir) = self.promise_idx_to_receipt_idx_with_sir(promise_idx)?;
 
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.delete_key_cost,
+            &self.fees_config.action_creation_config.delete_key_cost,
             sir,
         )?;
 
@@ -1114,7 +1120,7 @@ impl<'a> VMLogic<'a> {
         let (receipt_idx, sir) = self.promise_idx_to_receipt_idx_with_sir(promise_idx)?;
 
         self.gas_counter.pay_base_gas_fee(
-            &self.config.runtime_fees.action_creation_config.delete_account_cost,
+            &self.fees_config.action_creation_config.delete_account_cost,
             sir,
         )?;
 
@@ -1213,7 +1219,7 @@ impl<'a> VMLogic<'a> {
         let return_val = Self::memory_get(self.memory, value_ptr, value_len)?;
         let mut gas_use: Gas = 0;
         let num_bytes = return_val.len() as u64;
-        let data_cfg = &self.config.runtime_fees.data_receipt_creation_config;
+        let data_cfg = &self.fees_config.data_receipt_creation_config;
         for data_receiver in &self.context.output_data_receivers {
             let sir = data_receiver == &self.context.current_account_id;
             // We deduct for execution here too, because if we later have an OR combinator
@@ -1357,16 +1363,16 @@ impl<'a> VMLogic<'a> {
             registers,
             gas_counter,
             config,
+            fees_config,
             valid_iterators,
             invalid_iterators,
             ext,
             ..
         } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.storage_write_base)?;
+        gas_counter.pay_base(config.ext_costs.storage_write_base)?;
+        gas_counter.pay_per_byte(config.ext_costs.storage_write_key_byte, key_len)?;
         gas_counter
-            .pay_per_byte(config.runtime_fees.ext_costs.storage_write_key_byte, key_len)?;
-        gas_counter
-            .pay_per_byte(config.runtime_fees.ext_costs.storage_write_value_byte, value_len)?;
+            .pay_per_byte(config.ext_costs.storage_write_value_byte, value_len)?;
         // All iterators that were valid now become invalid
         for invalidated_iter_idx in valid_iterators.drain() {
             ext.storage_iter_drop(invalidated_iter_idx)?;
@@ -1375,7 +1381,7 @@ impl<'a> VMLogic<'a> {
         let key = Self::memory_get(*memory, key_ptr, key_len)?;
         let value = Self::memory_get(*memory, value_ptr, value_len)?;
         let evicted = self.ext.storage_set(&key, &value)?;
-        let storage_config = &config.runtime_fees.storage_usage_config;
+        let storage_config = &fees_config.storage_usage_config;
         match evicted {
             Some(old_value) => {
                 self.current_storage_usage -=
@@ -1412,14 +1418,14 @@ impl<'a> VMLogic<'a> {
     ///   `MemoryAccessViolation`.
     pub fn storage_read(&mut self, key_len: u64, key_ptr: u64, register_id: u64) -> Result<u64> {
         let Self { ext, memory, registers, gas_counter, config, .. } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.storage_write_base)?;
-        gas_counter.pay_per_byte(config.runtime_fees.ext_costs.storage_write_key_byte, key_len)?;
+        gas_counter.pay_base(config.ext_costs.storage_write_base)?;
+        gas_counter.pay_per_byte(config.ext_costs.storage_write_key_byte, key_len)?;
         let key = Self::memory_get(*memory, key_ptr, key_len)?;
         let read = ext.storage_get(&key)?;
         match read {
             Some(value) => {
                 gas_counter.pay_per_byte(
-                    config.runtime_fees.ext_costs.storage_write_value_byte,
+                    config.ext_costs.storage_write_value_byte,
                     value.len() as u64,
                 )?;
                 Self::internal_write_register(registers, gas_counter, config, register_id, &value)?;
@@ -1448,12 +1454,13 @@ impl<'a> VMLogic<'a> {
             registers,
             gas_counter,
             config,
+            fees_config,
             valid_iterators,
             invalid_iterators,
             ..
         } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.storage_remove_base)?;
-        gas_counter.pay_per_byte(config.runtime_fees.ext_costs.storage_remove_key_byte, key_len)?;
+        gas_counter.pay_base(config.ext_costs.storage_remove_base)?;
+        gas_counter.pay_per_byte(config.ext_costs.storage_remove_key_byte, key_len)?;
         // All iterators that were valid now become invalid
         for invalidated_iter_idx in valid_iterators.drain() {
             ext.storage_iter_drop(invalidated_iter_idx)?;
@@ -1461,11 +1468,11 @@ impl<'a> VMLogic<'a> {
         }
         let key = Self::memory_get(*memory, key_ptr, key_len)?;
         let removed = ext.storage_remove(&key)?;
-        let storage_config = &config.runtime_fees.storage_usage_config;
+        let storage_config = &fees_config.storage_usage_config;
         match removed {
             Some(value) => {
                 gas_counter.pay_per_byte(
-                    config.runtime_fees.ext_costs.storage_remove_ret_value_byte,
+                    config.ext_costs.storage_remove_ret_value_byte,
                     value.len() as u64,
                 )?;
                 self.current_storage_usage -=
@@ -1487,9 +1494,9 @@ impl<'a> VMLogic<'a> {
     ///
     /// If `key_len + key_ptr` exceeds the memory container it returns `MemoryAccessViolation`.
     pub fn storage_has_key(&mut self, key_len: u64, key_ptr: u64) -> Result<u64> {
-        self.gas_counter.pay_base(self.config.runtime_fees.ext_costs.storage_has_key_base)?;
+        self.gas_counter.pay_base(self.config.ext_costs.storage_has_key_base)?;
         self.gas_counter
-            .pay_per_byte(self.config.runtime_fees.ext_costs.storage_has_key_byte, key_len)?;
+            .pay_per_byte(self.config.ext_costs.storage_has_key_byte, key_len)?;
         let key = Self::memory_get(self.memory, key_ptr, key_len)?;
         let res = self.ext.storage_has_key(&key)?;
         Ok(res as u64)
@@ -1506,9 +1513,9 @@ impl<'a> VMLogic<'a> {
     /// If `prefix_len + prefix_ptr` exceeds the memory container it returns `MemoryAccessViolation`.
     pub fn storage_iter_prefix(&mut self, prefix_len: u64, prefix_ptr: u64) -> Result<u64> {
         self.gas_counter
-            .pay_base(self.config.runtime_fees.ext_costs.storage_iter_create_prefix_base)?;
+            .pay_base(self.config.ext_costs.storage_iter_create_prefix_base)?;
         self.gas_counter.pay_per_byte(
-            self.config.runtime_fees.ext_costs.storage_iter_create_key_byte,
+            self.config.ext_costs.storage_iter_create_key_byte,
             prefix_len,
         )?;
         let prefix = Self::memory_get(self.memory, prefix_ptr, prefix_len)?;
@@ -1534,15 +1541,15 @@ impl<'a> VMLogic<'a> {
         end_ptr: u64,
     ) -> Result<u64> {
         self.gas_counter
-            .pay_base(self.config.runtime_fees.ext_costs.storage_iter_create_range_base)?;
+            .pay_base(self.config.ext_costs.storage_iter_create_range_base)?;
         let start_key = Self::memory_get(self.memory, start_ptr, start_len)?;
         let end_key = Self::memory_get(self.memory, end_ptr, end_len)?;
         self.gas_counter.pay_per_byte(
-            self.config.runtime_fees.ext_costs.storage_iter_create_key_byte,
+            self.config.ext_costs.storage_iter_create_key_byte,
             start_len,
         )?;
         self.gas_counter.pay_per_byte(
-            self.config.runtime_fees.ext_costs.storage_iter_create_key_byte,
+            self.config.ext_costs.storage_iter_create_key_byte,
             end_len,
         )?;
         let iterator_index = self.ext.storage_iter_range(&start_key, &end_key)?;
@@ -1581,7 +1588,7 @@ impl<'a> VMLogic<'a> {
         let Self {
             ext, registers, gas_counter, config, valid_iterators, invalid_iterators, ..
         } = self;
-        gas_counter.pay_base(config.runtime_fees.ext_costs.storage_iter_next_base)?;
+        gas_counter.pay_base(config.ext_costs.storage_iter_next_base)?;
         if invalid_iterators.contains(&iterator_id) {
             return Err(HostError::IteratorWasInvalidated.into());
         } else if !valid_iterators.contains(&iterator_id) {
@@ -1592,11 +1599,11 @@ impl<'a> VMLogic<'a> {
         match value {
             Some((key, value)) => {
                 gas_counter.pay_per_byte(
-                    config.runtime_fees.ext_costs.storage_iter_next_key_byte,
+                    config.ext_costs.storage_iter_next_key_byte,
                     key.len() as u64,
                 )?;
                 gas_counter.pay_per_byte(
-                    config.runtime_fees.ext_costs.storage_iter_next_value_byte,
+                    config.ext_costs.storage_iter_next_value_byte,
                     value.len() as u64,
                 )?;
                 Self::internal_write_register(
