+++ runtime/near-vm-logic/tests/test_promises.rs
@@ -7,6 +7,7 @@ use near_vm_logic::mocks::mock_external::MockedExternal;
 use near_vm_logic::mocks::mock_memory::MockedMemory;
 use near_vm_logic::types::PromiseResult;
 use near_vm_logic::{Config, VMLogic};
+use near_runtime_fees::RuntimeFeesConfig;
 use serde_json;
 
 #[test]
@@ -14,6 +15,7 @@ fn test_promise_results() {
     let mut ext = MockedExternal::default();
     let context = get_context(vec![], false);
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
 
     let mut promise_results = vec![];
     promise_results.push(PromiseResult::Successful(b"test".to_vec()));
@@ -21,7 +23,7 @@ fn test_promise_results() {
     promise_results.push(PromiseResult::NotReady);
 
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
 
     assert_eq!(logic.promise_results_count(), Ok(3), "Total count of registers must be 3");
     assert_eq!(logic.promise_result(0, 0), Ok(1), "Must return code 1 on success");
@@ -38,9 +40,10 @@ fn test_promise_batch_action_function_call() {
     let mut ext = MockedExternal::default();
     let context = get_context(vec![], false);
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     let index = promise_create(&mut logic, b"rick.test", 0, 0).expect("should create a promise");
 
     promise_batch_action_function_call(&mut logic, 123, 0, 0)
@@ -78,9 +81,10 @@ fn test_promise_batch_action_create_account() {
     let mut ext = MockedExternal::default();
     let context = get_context(vec![], false);
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     let index = promise_create(&mut logic, b"rick.test", 0, 0).expect("should create a promise");
 
     logic
@@ -124,9 +128,10 @@ fn test_promise_batch_action_deploy_contract() {
     let mut ext = MockedExternal::default();
     let context = get_context(vec![], false);
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     let index = promise_create(&mut logic, b"rick.test", 0, 0).expect("should create a promise");
     let code = b"sample";
 
@@ -181,9 +186,10 @@ fn test_promise_batch_action_transfer() {
     context.account_balance = 100;
     context.attached_deposit = 10;
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     let index = promise_create(&mut logic, b"rick.test", 0, 0).expect("should create a promise");
 
     logic
@@ -237,9 +243,10 @@ fn test_promise_batch_action_stake() {
     let mut context = get_context(vec![], false);
     context.account_balance = 100;
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     let index = promise_create(&mut logic, b"rick.test", 0, 0).expect("should create a promise");
     let key = b"ed25519:5do5nkAEVhL8iteDvXNgxi4pWK78Y7DDadX11ArFNyrf";
 
@@ -314,9 +321,10 @@ fn test_promise_batch_action_add_key_with_function_call() {
     let mut context = get_context(vec![], false);
     context.account_balance = 100;
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     let index = promise_create(&mut logic, b"rick.test", 0, 0).expect("should create a promise");
     let key = b"ed25519:5do5nkAEVhL8iteDvXNgxi4pWK78Y7DDadX11ArFNyrf";
     let nonce = 1;
@@ -400,9 +408,10 @@ fn test_promise_batch_then() {
     let mut context = get_context(vec![], false);
     context.account_balance = 100;
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
 
     let account_id = b"rick.test";
     let index = promise_create(&mut logic, account_id, 0, 0).expect("should create a promise");
