+++ runtime/near-vm-logic/tests/test_view_method.rs
@@ -1,4 +1,5 @@
 use crate::fixtures::get_context;
+use near_runtime_fees::RuntimeFeesConfig;
 use near_vm_logic::mocks::mock_external::MockedExternal;
 use near_vm_logic::mocks::mock_memory::MockedMemory;
 use near_vm_logic::{Config, VMLogic};
@@ -10,9 +11,10 @@ macro_rules! test_prohibited {
         let mut ext = MockedExternal::default();
         let context = get_context(vec![], true);
         let config = Config::default();
+        let fees = RuntimeFeesConfig::default();
         let promise_results = vec![];
         let mut memory = MockedMemory::default();
-        let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+        let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
 
         let name = stringify!($f);
         logic.$f($($arg, )*).expect_err(&format!("{} is not allowed in view calls", name))
@@ -52,8 +54,9 @@ fn test_allowed_view_method() {
     let context = get_context(vec![], true);
     let block_index = context.block_index;
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let promise_results = vec![];
     let mut memory = MockedMemory::default();
-    let mut logic = VMLogic::new(&mut ext, context, &config, &promise_results, &mut memory);
+    let mut logic = VMLogic::new(&mut ext, context, &config, &fees, &promise_results, &mut memory);
     assert_eq!(logic.block_index().unwrap(), block_index);
 }
