+++ runtime/near-vm-runner-standalone/Cargo.toml
@@ -21,5 +21,6 @@ to make sure it has expected behavior once deployed to the blockchain.
 [dependencies]
 serde_json = "1.0"
 clap = "2.33.0"
-near-vm-logic = { path = "../near-vm-logic", features = ["mocks"], version = "0.3.0"}
-near-vm-runner = { path = "../near-vm-runner", version = "0.3.0" }
+near-vm-logic = { path = "../near-vm-logic", features = ["mocks"], version = "0.4.0"}
+near-vm-runner = { path = "../near-vm-runner", version = "0.4.0" }
+near-runtime-fees = { path = "../near-runtime-fees", version = "0.4.0" }
