+++ runtime/near-vm-runner-standalone/src/main.rs
@@ -9,6 +9,7 @@ use clap::{App, Arg};
 use near_vm_logic::mocks::mock_external::MockedExternal;
 use near_vm_logic::types::PromiseResult;
 use near_vm_logic::{Config, VMContext};
+use near_runtime_fees::RuntimeFeesConfig;
 use near_vm_runner::run;
 use std::fs;
 
@@ -119,8 +120,9 @@ fn main() {
         fs::read(matches.value_of("wasm-file").expect("Wasm file needs to be specified")).unwrap();
 
     let mut fake_external = MockedExternal::new();
+    let fees = RuntimeFeesConfig::default();
     let (outcome, err) =
-        run(vec![], &code, &method_name, &mut fake_external, context, &config, &promise_results);
+        run(vec![], &code, &method_name, &mut fake_external, context, &config, &fees, &promise_results);
 
     if let Some(outcome) = outcome {
         let str = serde_json::to_string(&outcome).unwrap();
