+++ runtime/near-vm-runner/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "near-vm-runner"
-version = "0.3.0"
+version = "0.4.0"
 authors = ["Near Inc <hello@nearprotocol.com>"]
 edition = "2018"
 license = "Apache-2.0"
@@ -16,13 +16,14 @@ This crate implements the specification of the interface that Near blockchain ex
 cached = "0.9.0"
 wasmer-runtime = { version = "0.7.0", features = ["singlepass"] }
 wasmer-runtime-core = { version = "0.7.0"}
-near-vm-logic = { path="../near-vm-logic", version = "0.3.0"}
-near-vm-errors = { path = "../near-vm-errors", version = "0.3.0" }
+near-runtime-fees = { path="../near-runtime-fees", version = "0.4.0" }
+near-vm-logic = { path="../near-vm-logic", version = "0.4.0"}
+near-vm-errors = { path = "../near-vm-errors", version = "0.4.0" }
 pwasm-utils = "0.11.0"
 parity-wasm = "0.40.1"
 
 [dev-dependencies]
-near-vm-logic = { path="../near-vm-logic", features=["mocks"], version = "0.3.0"}
+near-vm-logic = { path="../near-vm-logic", features=["mocks"], version = "0.4.0"}
 assert_matches = "1.3.0"
 wabt = "0.7.4"
 bencher = "0.1.5"
