+++ runtime/near-vm-runner/src/runner.rs
@@ -3,6 +3,7 @@ use std::ffi::c_void;
 use crate::errors::IntoVMError;
 use crate::memory::WasmerMemory;
 use crate::{cache, imports};
+use near_runtime_fees::RuntimeFeesConfig;
 use near_vm_errors::{FunctionCallError, MethodResolveError, VMError};
 use near_vm_logic::types::PromiseResult;
 use near_vm_logic::{Config, External, VMContext, VMLogic, VMOutcome};
@@ -34,7 +35,8 @@ pub fn run<'a>(
     method_name: &[u8],
     ext: &mut dyn External,
     context: VMContext,
-    config: &'a Config,
+    wasm_config: &'a Config,
+    fees_config: &'a RuntimeFeesConfig,
     promise_results: &'a [PromiseResult],
 ) -> (Option<VMOutcome>, Option<VMError>) {
     if method_name.is_empty() {
@@ -46,17 +48,17 @@ pub fn run<'a>(
         );
     }
 
-    let module = match cache::compile_cached_module(code_hash, code, config) {
+    let module = match cache::compile_cached_module(code_hash, code, wasm_config) {
         Ok(x) => x,
         Err(err) => return (None, Some(err)),
     };
-    let mut memory = match WasmerMemory::new(config) {
+    let mut memory = match WasmerMemory::new(wasm_config) {
         Ok(x) => x,
         Err(_err) => panic!("Cannot create memory for a contract call"),
     };
     let memory_copy = memory.clone();
 
-    let mut logic = VMLogic::new(ext, context, config, promise_results, &mut memory);
+    let mut logic = VMLogic::new(ext, context, wasm_config, fees_config, promise_results, &mut memory);
 
     let raw_ptr = &mut logic as *mut _ as *mut c_void;
     let import_object = imports::build(memory_copy, raw_ptr);
