+++ runtime/near-vm-runner/tests/test_rs_contract.rs
@@ -3,6 +3,7 @@ use near_vm_logic::mocks::mock_external::MockedExternal;
 use near_vm_logic::types::ReturnData;
 use near_vm_logic::{Config, VMContext, VMOutcome};
 use near_vm_runner::{run, VMError};
+use near_runtime_fees::RuntimeFeesConfig;
 use std::fs;
 use std::mem::size_of;
 use std::path::PathBuf;
@@ -53,6 +54,7 @@ pub fn test_read_write() {
 
     let context = create_context(&arr_u64_to_u8(&[10u64, 20u64]));
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
 
     let promise_results = vec![];
     let result = run(
@@ -62,13 +64,14 @@ pub fn test_read_write() {
         &mut fake_external,
         context,
         &config,
+        &fees,
         &promise_results,
     );
     assert_run_result(result, 0);
 
     let context = create_context(&arr_u64_to_u8(&[10u64]));
     let result =
-        run(vec![], &code, b"read_value", &mut fake_external, context, &config, &promise_results);
+        run(vec![], &code, b"read_value", &mut fake_external, context, &config, &fees, &promise_results);
     assert_run_result(result, 20);
 }
 
@@ -93,10 +96,11 @@ fn run_test_ext(method: &[u8], expected: &[u8], input: &[u8]) {
     let code = fs::read(path).unwrap();
     let mut fake_external = MockedExternal::new();
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
     let context = create_context(&input);
 
     let (outcome, err) =
-        run(input.to_owned(), &code, &method, &mut fake_external, context, &config, &[]);
+        run(input.to_owned(), &code, &method, &mut fake_external, context, &config, &fees, &[]);
 
     if let Some(_) = err {
         panic!("Failed execution: {:?}", err);
@@ -153,6 +157,7 @@ pub fn test_out_of_memory() {
 
     let context = create_context(&[]);
     let config = Config::default();
+    let fees = RuntimeFeesConfig::default();
 
     let promise_results = vec![];
     let result = run(
@@ -162,6 +167,7 @@ pub fn test_out_of_memory() {
         &mut fake_external,
         context,
         &config,
+        &fees,
         &promise_results,
     );
     assert_eq!(
