+++ runtime/runtime/src/actions.rs
@@ -169,6 +169,7 @@ pub(crate) fn action_function_call(
         &mut runtime_ext,
         context,
         &config.wasm_config,
+        &config.transaction_costs,
         promise_results,
     );
     if let Some(err) = err {
