+++ runtime/runtime/src/state_viewer.rs
@@ -13,6 +13,7 @@ use near_primitives::utils::{is_valid_account_id, prefix_for_data};
 use near_primitives::views::ViewStateResult;
 use near_store::{get_access_key, get_account, TrieUpdate};
 use near_vm_logic::{Config, ReturnData, VMContext};
+use near_runtime_fees::RuntimeFeesConfig;
 
 use crate::actions::get_code_with_cache;
 use crate::ethereum::EthashProvider;
@@ -134,6 +135,7 @@ impl TrieViewer {
                 &mut runtime_ext,
                 context,
                 &Config::default(),
+                &RuntimeFeesConfig::default(),
                 &[],
             )
         };
